// Marketeer.UI.Spa/src/app/models/models.d.ts
//     This code was generated by a Reinforced.Typings tool.
//     Changes to this file may cause incorrect behavior and will be lost if
//     the code is regenerated.

export interface IPaginateFilterDto
{
	pageIndex?: number;
	pageItemCount?: number;
	orderBy?: string;
	isOrderAsc?: boolean;
	isPaginated?: boolean;
}
export interface IStringDataDto
{
	data?: string;
}
export interface IClientTokenDto
{
	userId?: number;
	userName?: string;
	exp?: Date;
	iat?: Date;
	nbf?: Date;
	roles?: string[];
}
export interface ILoginDto
{
	userName?: string;
	password?: string;
}
export enum LogLevel {
	Trace = 0,
	Debug = 1,
	Information = 2,
	Warning = 3,
	Error = 4,
	Critical = 5,
	None = 6
}